refactor: 예외 처리 시 AccountException을 사용하도록 변경

+++ src/main/java/com/example/account/repository/AccountRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 @Repository
 public interface AccountRepository extends JpaRepository<Account, Long> {
-    long countByAccountUser(AccountUser user);
+    Integer countByAccountUser(AccountUser user);
     Optional<Account> findByAccountNumber(String accountNumber);
     Optional<Account> findFirstByOrderByIdDesc();
 }
